Add XForwardedHeaderSupport
Should help fix the wrong redirect URL being created on Heroku. The docs
say that X-Forwarded-* headers cannot be trusted, but to the best of my
knowledge, because Discord allows you to limit acceptable redirect URLs
even if somebody was to somehow send their own header, Discord would
reject it for not returning to the proper set list. They also wouldn't
have the client secret.