Remove FA thumbnail option
The feature is rarely used and those who do use it were not disabling
thumbnails, they were disabling the feature entirely. This is the first
steps in reducing the total number of config options for eventually
allowing editing directly within the bot. (Making a website was a fun
exercise, but far from ideal for a number of reasons.)