Remember existing starboard messages to update them
Webhook messages can be edited, we can now add a count that updates.
Redis will allow us to remember the "link" between an original message
and it on the starboard. This code hasn't been tested yet though.

I really hate the WebhookDirector and my edits were trying to be the
bare minimum to get this working. Not fixing it currently.