Do not allowed editing messages after being sent to the starboard
It seems weird to preserve history on the starboard, but we don't want
to make it a place you get once and then can edit what people thought
was worth.

Consider the following...
Person A: writes funny message
Person B: stars the message, sending it to the board
Person A: changes the complete meaning of their message
Person B: is insulted their "endorsement" has been defiled

Instead, once you make it to the board, you're stuck with what it was
that made it worth getting there. Deleted messages however are
permanently removed because sometimes things deserve to be forgotten.

There is a potential race condition between an update and "punishment".