Starboard 2 (#45)
* Begin simplify starboard preprocessing

I really hate how a lot of this is written, especially interacting with
webhooks.

* Remember existing starboard messages to update them

Webhook messages can be edited, we can now add a count that updates.
Redis will allow us to remember the "link" between an original message
and it on the starboard. This code hasn't been tested yet though.

I really hate the WebhookDirector and my edits were trying to be the
bare minimum to get this working. Not fixing it currently.

* Fix logic mistake and race condition

* Handle reaction removal and try alternate count location

* Sort of recover from failure to send first starboard message

If the bot dies before deleting, what happens?

* Implement Redlock for first send as well as expiring starboard tracking

* Fix Redlock unlock and starboard message creation criteria

* Update JDA and only retry on missing message

If the webhook is missing, it's your fault for deleting it. I can't be
responsible for fixing all of your mistakes. Just wait a bit for it to
work itself out when it expires.

* Attempt to delete starboard messages for deleted message

* Suppress WebhookClient logs

They're loud and verbose for something I only need one line of info
about

* Stop wastefully checking on stars from bots

* Respectfully check embed size limits

* Do not allowed editing messages after being sent to the starboard

It seems weird to preserve history on the starboard, but we don't want
to make it a place you get once and then can edit what people thought
was worth.

Consider the following...
Person A: writes funny message
Person B: stars the message, sending it to the board
Person A: changes the complete meaning of their message
Person B: is insulted their "endorsement" has been defiled

Instead, once you make it to the board, you're stuck with what it was
that made it worth getting there. Deleted messages however are
permanently removed because sometimes things deserve to be forgotten.

There is a potential race condition between an update and "punishment".

* Do not create a message on editing a deleted starboard message

* Mark deleted messages instead of mysteriously deleting

Dead messages no longer show the original author out of respect for the
dead.